+++ api/cmd/kubermatic-api/options.go
@@ -29,6 +29,7 @@ type serverRunOptions struct {
 	domain             string
 	exposeStrategy     corev1.ServiceType
 	dynamicDatacenters bool
+	dynamicPresets     bool
 	namespace          string
 	log                kubermaticlog.Options
 	accessibleAddons   sets.String
@@ -87,6 +88,7 @@ func newServerRunOptions() (serverRunOptions, error) {
 	flag.StringVar(&s.serviceAccountSigningKey, "service-account-signing-key", "", "Signing key authenticates the service account's token value using HMAC. It is recommended to use a key with 32 bytes or longer.")
 	flag.StringVar(&rawExposeStrategy, "expose-strategy", "NodePort", "The strategy to expose the controlplane with, either \"NodePort\" which creates NodePorts with a \"nodeport-proxy.k8s.io/expose: true\" annotation or \"LoadBalancer\", which creates a LoadBalancer")
 	flag.BoolVar(&s.dynamicDatacenters, "dynamic-datacenters", false, "Whether to enable dynamic datacenters")
+	flag.BoolVar(&s.dynamicPresets, "dynamic-presets", false, "Whether to enable dynamic presets")
 	flag.StringVar(&s.namespace, "namespace", "kubermatic", "The namespace kubermatic runs in, uses to determine where to look for datacenter custom resources")
 	flag.Parse()
 
@@ -156,4 +158,5 @@ type providers struct {
 	userInfoGetter                        provider.UserInfoGetter
 	settingsProvider                      provider.SettingsProvider
 	adminProvider                         provider.AdminProvider
+	presetProvider                        provider.PresetProvider
 }
