+++ api/pkg/handler/v1/common/common.go
@@ -61,13 +61,6 @@ type UpdateManager interface {
 	GetPossibleUpdates(from, clusterType string) ([]*version.Version, error)
 }
 
-// PresetsManager specifies a set of methods to handle presets for specific provider
-type PresetsManager interface {
-	GetPresets(userInfo *provider.UserInfo) ([]kubermaticv1.Preset, error)
-	GetPreset(userInfo *provider.UserInfo, name string) (*kubermaticv1.Preset, error)
-	SetCloudCredentials(userInfo *provider.UserInfo, presetName string, cloud kubermaticv1.CloudSpec, dc *kubermaticv1.Datacenter) (*kubermaticv1.CloudSpec, error)
-}
-
 // ServerMetrics defines metrics used by the API.
 type ServerMetrics struct {
 	HTTPRequestsTotal          *prometheus.CounterVec
