+++ api/pkg/handler/v1/provider/digitalocean.go
@@ -57,7 +57,7 @@ func DigitaloceanSizeWithClusterCredentialsEndpoint(projectProvider provider.Pro
 	}
 }
 
-func DigitaloceanSizeEndpoint(credentialManager common.PresetsManager, userInfoGetter provider.UserInfoGetter) endpoint.Endpoint {
+func DigitaloceanSizeEndpoint(presetsProvider provider.PresetProvider, userInfoGetter provider.UserInfoGetter) endpoint.Endpoint {
 	return func(ctx context.Context, request interface{}) (interface{}, error) {
 		req := request.(DoSizesReq)
 
@@ -68,7 +68,7 @@ func DigitaloceanSizeEndpoint(credentialManager common.PresetsManager, userInfoG
 		}
 
 		if len(req.Credential) > 0 {
-			preset, err := credentialManager.GetPreset(userInfo, req.Credential)
+			preset, err := presetsProvider.GetPreset(userInfo, req.Credential)
 			if err != nil {
 				return nil, errors.New(http.StatusInternalServerError, fmt.Sprintf("can not get preset %s for user %s", req.Credential, userInfo.Email))
 			}
