+++ api/pkg/leaderelection/leaderelection.go
@@ -23,7 +23,7 @@ const (
 	retryPeriod   = 2 * time.Second
 )
 
-// New returns a new leader elector which uses the "hostname + name" as lock identity
+// NewPresetsProvider returns a new leader elector which uses the "hostname + name" as lock identity
 func New(name string, leaderElectionClient kubernetes.Interface, recorder resourcelock.EventRecorder, callbacks leaderelection.LeaderCallbacks) (*leaderelection.LeaderElector, error) {
 	// Identity used to distinguish between multiple controller manager instances
 	id, err := os.Hostname()
