+++ api/pkg/provider/cloud/azure/provider.go
@@ -52,7 +52,7 @@ type Azure struct {
 	secretKeySelector provider.SecretKeySelectorValueFunc
 }
 
-// New returns a new Azure provider.
+// NewPresetsProvider returns a new Azure provider.
 func New(dc *kubermaticv1.Datacenter, secretKeyGetter provider.SecretKeySelectorValueFunc) (*Azure, error) {
 	if dc.Spec.Azure == nil {
 		return nil, errors.New("datacenter is not an Azure datacenter")
