+++ api/pkg/provider/types.go
@@ -504,3 +504,10 @@ type AdminProvider interface {
 	SetAdmin(userInfo *UserInfo, email string, isAdmin bool) (*kubermaticv1.User, error)
 	GetAdmins(userInfo *UserInfo) ([]kubermaticv1.User, error)
 }
+
+// PresetProvider declares the set of methods for interacting with presets
+type PresetProvider interface {
+	GetPresets(userInfo *UserInfo) ([]kubermaticv1.Preset, error)
+	GetPreset(userInfo *UserInfo, name string) (*kubermaticv1.Preset, error)
+	SetCloudCredentials(userInfo *UserInfo, presetName string, cloud kubermaticv1.CloudSpec, dc *kubermaticv1.Datacenter) (*kubermaticv1.CloudSpec, error)
+}
