+++ config/kubermatic/crd/crd-presets.yaml
@@ -1,0 +1,13 @@
+apiVersion: apiextensions.k8s.io/v1beta1
+kind: CustomResourceDefinition
+metadata:
+  name: presets.kubermatic.k8s.io
+spec:
+  group: kubermatic.k8s.io
+  names:
+    kind: Preset
+    listKind: PresetList
+    plural: presets
+    singular: preset
+  scope: Cluster
+  version: v1
